1.Use popups to show messages to the users for confirmation 2.Use a loading message together with a spinner to show the user while the app is downloading information from a server 3.Use gestures within your app.

+++ hookipaMobile/www/js/controllers.js
@@ -275,19 +275,31 @@ angular.module('hookipaMobile.controllers', [])
             
                     }])
 
-        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
-
-    $scope.baseURL = baseURL;
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', 
+    '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, 
+    favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
+    
+         $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
+
+    $ionicLoading.show({
+        template: '<ion-spinner></ion-spinner> Loading...'
+    });
 
     $scope.favorites = favoriteFactory.getFavorites();
 
     $scope.dishes = menuFactory.getDishes().query(
         function (response) {
             $scope.dishes = response;
+            $timeout(function () {
+                $ionicLoading.hide();
+            }, 1000);
         },
         function (response) {
             $scope.message = "Error: " + response.status + " " + response.statusText;
+            $timeout(function () {
+                $ionicLoading.hide();
+            }, 1000);
         });
     console.log($scope.dishes, $scope.favorites);
 
@@ -296,12 +308,30 @@ angular.module('hookipaMobile.controllers', [])
         console.log($scope.shouldShowDelete);
     }
 
-    $scope.deleteFavorite = function (index) {
-        
-        favoriteFactory.deleteFromFavorites(index);
+        $scope.deleteFavorite = function (index) {
+
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Confirm Delete',
+            template: 'Are you sure you want to delete this item?'
+        });
+
+        confirmPopup.then(function (res) {
+            if (res) {
+                console.log('Ok to delete');
+                favoriteFactory.deleteFromFavorites(index);
+            } else {
+                console.log('Canceled delete');
+            }
+        });
+
         $scope.shouldShowDelete = false;
 
-    }}])
+    }
+
+
+
+
+    }])
 
 
 .filter('favoriteFilter', function () {
@@ -315,7 +345,7 @@ angular.module('hookipaMobile.controllers', [])
         }
         return out;
 
-    }});
+    }})
